Create README.md
The dataset includes information about individual transactions between January 2018 and September 2019. A csv file containing transaction data and an xls file containing the budget for each transaction category made up our two files. Using power query, the data was cleaned and converted. For simplicity of calculation, a separate date table and monthly budget table are established using DAX. In a table, all the measures are kept in their appropriate display folders. 

The report's delicate and informative style provides us with the following details:
The financial analysis page offers thorough details on net income by month, year-to-date income and expense, and comparisons to the prior year's income and expense. The page for budget analysis includes information on the expenses incurred in relation to the allocated budget for each transaction type as well as the leftover budget. 

I welcome constructive feedback and your insightful recommendations.